requirements.txt
aniso8601==3.0.2
botocore==1.12.16
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.9
docutils==0.14
Flask==1.0.2
flask-restplus==0.12.1
future==0.16.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
line-bot-sdk==1.8.0
MarkupSafe==1.0
marshmallow==2.15.6
pyasn1==0.4.4
pymongo==3.7.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
